Move ZstdCompression mode out of Zstd814StoredFieldsFormat
